refactor(scm): move scm types and event types to scm
