+++ src/dev/cdevents/types.ts
@@ -8,7 +8,7 @@ import { VocabularySchema } from "../cdevents/schemas"
 // https://github.com/package-url/purl-spec/blob/master/PURL-SPECIFICATION.rst
 const componentPattern = /^[a-zA-Z0-9._-]+$/
 const urlWithProtocolPattern = /^[a-zA-Z][a-zA-Z0-9+.-]*:\/\/[^\s/$.?#].[^\s]*$/;
-const urlSchema = z.string().regex(urlWithProtocolPattern, {
+export const urlSchema = z.string().regex(urlWithProtocolPattern, {
     message: "Invalid URL format. A protocol (e.g., https, git) is required.",
 });
 const typeSchema = z.string().regex(componentPattern { message: "Invalid type format" });
@@ -99,94 +99,6 @@ const TaskRunFinishedEventSchema = z.object({
     errors: z.string().optional()
 });
 
-/*
-    Repository.
-    An SCM repository is identified by a name, an owner which can be a user or an organization, a url which is where the repository is hosted and optionally a viewUrl, which is a web location for humans to browse the content of the repository.
-*/
-const RepositorySchema = z.object({
-    id: z.string().uuid(),
-    source: z.string(),
-    type: z.literal("repository"),
-    name: z.string(),
-    owner: z.string(),
-    url: urlSchema, // url for API operations. Needs to include protocol
-    viewUrl: z.string().url()
-})
-
-/* SCM branch. Identified by ID
-*/
-const BranchSchema = z.object({
-    id: z.string(),
-    source: z.string(),
-    type: z.literal("branch"),
-    repository: RepositorySchema
-})
-
-/* SCM Change */
-const ChangeSchema = z.object({
-    id: z.string(),
-    source: z.string(),
-    type: z.literal("change"),
-    description: z.string(),
-    repository: RepositorySchema,
-})
-
-/* SCM Events */
-// const RepositoryCreatedEventSchema // <- in repository-events.ts
-// const RepositoryModifiedEventSchema // <- in repository-events.ts
-// const RepositoryDeletedEventSchema // <- in repository-events.ts
-
-// Branch Events
-const BranchCreatedEventSchema = z.object({
-    id: z.string(),
-    source: z.string().optional(),
-    type: VocabularySchema.optional(),
-    repository: RepositorySchema.optional()
-});
-
-const BranchDeletedEventSchema = z.object({
-    id: z.string(),
-    source: z.string().optional(),
-    type: VocabularySchema.optional(),
-    repository: RepositorySchema.optional()
-});
-
-// Change Events
-const ChangeCreatedEventSchema = z.object({
-    id: z.string(),
-    source: z.string().optional(),
-    type: VocabularySchema.optional(),
-    description: z.string(),
-    repository: RepositorySchema.optional()
-});
-
-const ChangeReviewedEventSchema = z.object({
-    id: z.string(),
-    source: z.string().optional(),
-    type: VocabularySchema.optional(),
-    repository: RepositorySchema
-});
-
-const ChangeMergedEventSchema = z.object({
-    id: z.string(),
-    source: z.string().optional(),
-    type: VocabularySchema.optional(),
-    repository: RepositorySchema.optional()
-});
-
-const ChangeAbandonedEventSchema = z.object({
-    id: z.string(),
-    source: z.string().optional(),
-    type: VocabularySchema.optional(),
-    repository: RepositorySchema.optional()
-});
-
-const ChangeUpdatedEventSchema = z.object({
-    id: z.string(),
-    source: z.string().optional(),
-    type: VocabularySchema.optional(),
-    repository: RepositorySchema.optional()
-})
 
 // Exported types
 // purl
@@ -205,17 +117,3 @@ export type PipelineRunStarted = z.infer<typeof PipelineRunStartedEventSchema>;
 export type PipelineFinished = z.infer<typeof PipelineRunFinishedEventSchema>;
 export type TaskRunStarted = z.infer<typeof TaskRunStartedEventSchema>;
 export type TaskRunFinished = z.infer<typeof TaskRunFinishedEventSchema>;
-
-/* SCM types */
-export type Repository = z.infer<typeof RepositorySchema>;
-export type Branch = z.infer<typeof BranchSchema>;
-export type Change = z.infer<typeof ChangeSchema>;
-
-/* SCM Event Types */
-export type BranchCreated = z.infer<typeof BranchCreatedEventSchema>;
-export type BranchDeleted = z.infer<typeof BranchDeletedEventSchema>;
-export type ChangeCreated = z.infer<typeof ChangeCreatedEventSchema>;
-export type ChangeReviewed = z.infer<typeof ChangeReviewedEventSchema>;
-export type ChangeMerged = z.infer<typeof ChangeMergedEventSchema>;
-export type ChangeAbandoned = z.infer<typeof ChangeAbandonedEventSchema>;
-export type ChangeUpdated = z.infer<typeof ChangeUpdatedEventSchema>;
